Only display the ACTUAL size in bytes -- no double counting circular references
Add a modified version of Base's summarysize, which computes a tree of
the sizes of each object, skipping objects that are already seen. This
is modified from Base, so that rather than simply _summing_ the sizes of
various fields, the sizes are collected into a tree structure which can
then be interactively traversed! :)