Add basic ability to export a PProf profile of memory usage
Uses the PProf package to export and open an interactive graph view of
how bytes are allocated by type within a given structure.

It still doesn't do a great job distinguishing between multiple
instances of the same type within a given parent, but it's pretty good.

----

I'm also not sure if it makes more sense to add this PProf integration
here, or to use MemoryInspector in PProf to provide a memory view from
PProf.jl. I think it makes more sense here?